refactor(workers): extract shared shell helpers to scripts/workers/shell.ts
- Single source of truth for runCommand, isToolAvailable, resolveRuff, and
  the Windows .cmd shim routing. Eliminates drift between format.worker
  and lint.worker (format was missing the Phase 2 Windows fix).
- resolveRuff now returns a readonly [string, ...string[]] non-empty tuple,
  so destructuring narrows the first element to string without needing the
  ! non-null assertion that violated repo conventions.
- format.worker.ts now also gets the Windows shim handling for free if
  any future format target uses npx.

Addresses my deep-review findings #1, #2 and Copilot review feedback.
